Code is reorganized in due to review
